Change the precision of size and uSize from mediumP to highP
When I checked the value of uSize and size set in the application, the
multiplication of these values could result in INF number if shader uses
mediump (fp16).

To avoid this problem, i change the precision of size and uSize from mediumP to highP

+++ dali-toolkit/internal/visuals/animated-gradient/animated-gradient-visual.cpp
@@ -96,7 +96,7 @@ const Toolkit::DevelAnimatedGradientVisual::AnimationParameter::EasingType::Type
 const char* const BASIC_VERTEX_SHADER = DALI_COMPOSE_SHADER(
   attribute mediump vec2 aPosition;
   uniform highp   mat4 uMvpMatrix;
-  uniform mediump vec3 uSize;
+  uniform highp   vec3 uSize;
 
   uniform mediump vec2 start_point;
   uniform mediump vec2 end_point;
@@ -118,7 +118,7 @@ const char* const BASIC_VERTEX_SHADER = DALI_COMPOSE_SHADER(
 
   //Visual size and offset
   uniform mediump vec2 offset;
-  uniform mediump vec2 size;
+  uniform highp vec2 size;
   uniform mediump vec4 offsetSizeMode;
   uniform mediump vec2 origin;
   uniform mediump vec2 anchorPoint;
